requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.6
lxml==4.2.1
Pillow==5.1.0
requests==2.18.4
urllib3==1.22
